trying to sort of file structure and adding in game constants and game objects
