+++ src/game/CMakeLists.txt
@@ -1,5 +1,7 @@
 set(GAME
-   src/game/Player.h
-   src/game/Player.c
+   src/game/GameConstants.h
+   src/game/GameConstants.c
+   src/game/GameObjects/Camera.h
+   src/game/GameObjects/Camera.c
 
     PARENT_SCOPE)
