+++ src/game/GameConstants.c
@@ -1,0 +1,4 @@
+#include "GameConstants.h"
+const float MOUSE_SENS = 0.001f;
+const float WALK_SPEED = 15.0f;
+const float UP_SPEED = 10.0f;
