+++ src/game/GameConstants.h
@@ -1,0 +1,5 @@
+#pragma once
+
+extern const float MOUSE_SENS;
+extern const float WALK_SPEED;
+extern const float UP_SPEED;
