+++ src/window/CMakeLists.txt
@@ -1,5 +1,7 @@
 set(WINDOW
    src/window/Window.h
    src/window/Window.c
+   src/window/Events.h
+   src/window/Events.c
     PARENT_SCOPE)
